+++ src/cli.rs
@@ -64,7 +64,8 @@ and may be extended with custom CLI options additionally.
 #     }
 # }
 #
-# let fut = async {
+# #[tokio::main(flavor = "current_thread")]
+# async fn main() {
 #[derive(StructOpt)]
 struct CustomOpts {
     /// Additional time to wait in before hook.
@@ -83,13 +84,7 @@ MyWorld::cucumber()
     .with_cli(opts)
     .run_and_exit("tests/features/readme")
     .await;
-# };
-#
-# tokio::runtime::Builder::new_current_thread()
-#    .enable_all()
-#    .build()
-#    .unwrap()
-#    .block_on(fut);
+# }
 ```
 
 [`Cucumber`]: crate::Cucumber
@@ -178,9 +173,7 @@ This struct is especially useful, when implementing custom [`Writer`] wrapping
 another one:
 ```rust
 # use async_trait::async_trait;
-# use cucumber::{
-#     cli, event, parser, ArbitraryWriter, Event, FailureWriter, World, Writer,
-# };
+# use cucumber::{cli, event, parser, writer, Event, World, Writer};
 # use structopt::StructOpt;
 #
 struct CustomWriter<Wr>(Wr);
@@ -213,11 +206,11 @@ where
 // Useful blanket impls:
 
 #[async_trait(?Send)]
-impl<'val, W, Wr, Val> ArbitraryWriter<'val, W, Val> for CustomWriter<Wr>
+impl<'val, W, Wr, Val> writer::Arbitrary<'val, W, Val> for CustomWriter<Wr>
 where
     W: World,
     Self: Writer<W>,
-    Wr: ArbitraryWriter<'val, W, Val>,
+    Wr: writer::Arbitrary<'val, W, Val>,
     Val: 'val,
 {
     async fn write(&mut self, val: Val)
@@ -228,11 +221,11 @@ where
     }
 }
 
-impl<W, Wr> FailureWriter<W> for CustomWriter<Wr>
+impl<W, Wr> writer::Failure<W> for CustomWriter<Wr>
 where
     W: World,
     Self: Writer<W>,
-    Wr: FailureWriter<W>,
+    Wr: writer::Failure<W>,
 {
     fn failed_steps(&self) -> usize {
         self.0.failed_steps()
@@ -246,6 +239,12 @@ where
         self.0.hook_errors()
     }
 }
+
+impl<Wr: writer::Normalized> writer::Normalized for CustomWriter<Wr> {}
+
+impl<Wr: writer::NonTransforming> writer::NonTransforming
+    for CustomWriter<Wr>
+{}
 ```
 
 [`Writer`]: crate::Writer