+++ CHANGELOG.md
@@ -11,15 +11,37 @@ All user visible changes to `cucumber` crate will be documented in this file. Th
 
 [Diff](/../../compare/v0.10.2...v0.11.0) | [Milestone](/../../milestone/3)
 
+### BC Breaks
+
+- Moved `World` type parameter of `WriterExt` trait to methods. ([#160])
+- Renamed `Normalized` and `Summarized` `Writer`s to `Normalize` and `Summarize`. ([#162])
+- Removed `writer::Basic` `Default` impl and change `writer::Basic::new()` return type to `writer::Normalize<writer::Basic>`. ([#162])
+
 ### Added
 
 - Ability for step functions to return `Result`. ([#151])
-- Arbitrary output for `writer::Basic` ([#147])
-- `writer::JUnit` ([JUnit XML report][0110-1]) behind the `output-junit` feature flag ([#147])
+- Arbitrary output for `writer::Basic`. ([#147])
+- `writer::JUnit` ([JUnit XML report][0110-1]) behind the `output-junit` feature flag. ([#147])
+- `writer::Json` ([Cucumber JSON format][0110-2]) behind the `output-json` feature flag. ([#159])
+- `writer::Tee` for outputting to multiple terminating `Writer`s simultaneously. ([#160])
+- `writer::discard::Arbitrary` and `writer::discard::Failure` for providing no-op implementations of the corresponding `Writer` traits. ([#160])
+- Inability to build invalid `Writer`s pipelines:
+    - `writer::Normalized` trait required for `Writer`s in `Cucumber` running methods. ([#162])
+    - `writer::NonTransforming` trait required for `writer::Repeat`. ([#162])
+    - `writer::Summarizable` trait required for `writer::Summarize`. ([#162])
+
+### Fixed
+
+- Template regex in `Scenario Outline` expansion from `<(\S+)>` to `<([^>\s]+)>`. ([#163])
 
 [#147]: /../../pull/147
 [#151]: /../../pull/151
+[#159]: /../../pull/159
+[#160]: /../../pull/160
+[#162]: /../../pull/162
+[#163]: /../../pull/163
 [0110-1]: https://llg.cubic.org/docs/junit
+[0110-2]: https://github.com/cucumber/cucumber-json-schema